requirements.txt
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
nltk==3.5
pipenv==2020.8.13
python-Levenshtein==0.12.0
regex==2020.7.14
requests==2.24.0
six==1.15.0
stop-words==2018.7.23
tqdm==4.48.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
Werkzeug==1.0.1
